backend/src/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jax==0.4.30
jaxlib==0.4.30
jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
starlette==0.37.2
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.42
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
sqlalchemy==2.0.31
pre-commit
python-magic==0.4.27
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
minio==7.2.7
pycryptodome==3.20.0
